requirements.txt
PyQt5
groq
python-dotenv